requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
daphne==4.0.0
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.5
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.5
packaging==23.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.2.0
python-decouple==3.8
pytz==2023.3
redis==4.6.0
requests==2.30.0
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.0
uritemplate==4.1.1
urllib3==2.0.2
zope.interface==6.0
